requirements.txt
ansicolors==1.1.8
ansiwrap==0.8.4
arabic-reshaper==2.1.3
autopep8==1.5.4
backports.csv==1.0.7
boto==2.49.0
boto3==1.5.0
botocore==1.8.14
cairocffi==0.9.0
CairoSVG==2.4.0
certifi==2019.3.9
cffi==1.14.4
chainmap==1.0.3
chardet==3.0.4
combomethod==1.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
cssselect2==0.2.1
decorator==4.4.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
django-admin-rangefilter==0.5.0
django-braces==1.14.0
django-cors-headers==3.1.0
django-cors-middleware==1.4.0
django-extensions==2.2.5
django-filter==2.2.0
django-google-maps==0.11.0
django-import-export==1.2.0
django-jet==1.0.8
django-map-widgets==0.2.2
django-mathfilters==0.4.0
django-nested-admin==3.2.4
django-nested-inline==0.3.7
django-nose==1.4.6
django-oauth-toolkit==1.3.3
django-pdfkit==0.3.1
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.1.1
django-ses==0.8.10
django-smart-selects==1.5.4
django-storages==1.6.3
django-summernote==0.8.11.4
django-super-inlines==0.1.4
djangorestframework==3.12.2
docopt==0.6.2
docutils==0.14
drf-social-oauth2==1.0.8
et-xmlfile==1.0.1
future==0.17.1
greenlet==0.4.15
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
imageio==2.5.0
imageio-ffmpeg==0.2.0
intspan==1.6.1
itypes==1.1.0
jdcal==1.4.1
jet-django==0.4.3
Jinja2==2.10
jmespath==0.9.4
Markdown==3.1.1
MarkupSafe==1.1.1
mementos==1.3.1
moviepy==1.0.0
mysqlclient==2.1.0
newrelic==3.2.2.94
nose==1.3.7
nulltype==2.3.1
num2words==0.5.10
numpy==1.19.4
oauthlib==3.1.0
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==2.6.2
options==1.4.10
pdfkit==0.6.0
Pillow==8.0.1
proglog==0.1.9
pycodestyle==2.6.0
pycparser==2.19
pyfcm==1.4.5
Pygments==2.6.1
PyJWT==1.7.1
PyMySQL==0.9.3
PyPDF2==1.26.0
Pyphen==0.9.5
python-bidi==0.4.2
python-dateutil==2.8.0
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2019.3
PyYAML==3.12
razorpay==1.2.0
reportlab==3.5.68
requests==2.21.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
s3transfer==0.1.13
say==1.6.6
simplejson==3.16.0
simplere==1.2.13
six==1.12.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
sqlparse==0.3.0
tablib==0.13.0
textdata==2.4.1
textwrap3==0.9.2
tinycss2==1.0.2
toml==0.10.2
tqdm==4.31.1
uritemplate==3.0.0
urllib3==1.24.1
WeasyPrint==47
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==1.2.0
xlwt==1.3.0
